==> notebooks/01_morph2-balanced.ipynb <==
import pandas as pd
import numpy as np

# Train

df_train = pd.read_csv('morph2_train.csv', index_col=0)
df_train.head()

bins = np.bincount(df_train['age'].values)
bins

bins[:33]

lst = []
for idx, i in enumerate(bins[:33]):
    tmp = df_train[df_train['age'] == idx]
    lst.append(tmp.iloc[:625, ])

df_train_new = pd.concat(lst)

df_train_new.head()

bins = np.bincount(df_train_new['age'].values)
bins

df_train_new.to_csv('morph2_train_balanced.csv')

# Valid

df_valid = pd.read_csv('morph2_valid.csv', index_col=0)
df_valid.head()

bins = np.bincount(df_valid['age'].values)
bins

bins[:33]

lst = []
for idx, i in enumerate(bins[:33]):
    tmp = df_valid[df_valid['age'] == idx]
    lst.append(tmp.iloc[:55, ])

df_valid_new = pd.concat(lst)

df_valid_new.head()

bins = np.bincount(df_valid_new['age'].values)
bins

df_valid_new.to_csv('morph2_valid_balanced.csv')

# Test

df_test = pd.read_csv('morph2_test.csv', index_col=0)
df_test.head()

bins = np.bincount(df_test['age'].values)
bins

bins[:33]

lst = []
for idx, i in enumerate(bins[:33]):
    tmp = df_test[df_test['age'] == idx]
    lst.append(tmp.iloc[:190, ])

df_test_new = pd.concat(lst)

df_test_new.head()

bins = np.bincount(df_test_new['age'].values)
bins

df_test_new.to_csv('morph2_test_balanced.csv')
