==> notebooks/04_afad-splits.ipynb <==
import pandas as pd

df_train = pd.read_csv('old_afad_train.csv')
df_train.head()

df_valid = pd.read_csv('old_afad_valid.csv')
df_valid.head()

df_test = pd.read_csv('old_afad_test.csv')
df_test.head()

df = pd.concat((df_train, df_valid, df_test))

df.to_csv('afad.csv', index=None)

# Split

import numpy as np

np.bincount(df['age'].values)

indices = np.arange(df.shape[0])
labels = df['age'].values

from sklearn.model_selection import train_test_split


indices_train, indices_test, labels_train, labels_test = train_test_split(
    indices, labels, test_size=0.2, random_state=123, shuffle=True, stratify=labels)

indices_train, indices_valid, labels_train, labels_valid = train_test_split(
    indices_train, labels_train, test_size=0.05, random_state=123, shuffle=True, stratify=labels_train)

df_train = df.iloc[indices_train]
df_train.to_csv('afad_train.csv', index=None)

df_valid = df.iloc[indices_valid]
df_valid.to_csv('afad_valid.csv', index=None)

df_test = df.iloc[indices_test]
df_test.to_csv('afad_test.csv', index=None)
