==> notebooks/05_afad-balanced.ipynb <==
import pandas as pd
import numpy as np

# Train

df_train = pd.read_csv('afad_train.csv')
df_train.head()

bins = np.bincount(df_train['age'].values)
bins

bins[3:16]

lst = []
for idx, i in enumerate(bins[3:16]):
    tmp = df_train[df_train['age'] == idx+3]
    lst.append(tmp.iloc[:3550, ])

df_train_new = pd.concat(lst)

df_train_new.head()

df_train_new['age'] = df_train_new['age'] - 3

bins = np.bincount(df_train_new['age'].values)
bins

df_train_new.to_csv('afad_train_balanced.csv', index=None)

# Valid

df_valid = pd.read_csv('afad_valid.csv')
df_valid.head()

bins = np.bincount(df_valid['age'].values)
bins

bins[3:16]

lst = []
for idx, i in enumerate(bins[3:16]):
    tmp = df_valid[df_valid['age'] == idx+3]
    lst.append(tmp.iloc[:189, ])

df_valid_new = pd.concat(lst)

df_valid_new.head()

df_valid_new['age'] = df_valid_new['age'] - 3

bins = np.bincount(df_valid_new['age'].values)
bins

df_valid_new.to_csv('afad_valid_balanced.csv', index=None)

# Test

df_test = pd.read_csv('afad_test.csv')
df_test.head()

bins = np.bincount(df_test['age'].values)
bins

bins[3:16]

lst = []
for idx, i in enumerate(bins[3:16]):
    tmp = df_test[df_test['age'] == idx+3]
    lst.append(tmp.iloc[:945, ])

df_test_new = pd.concat(lst)

df_test_new.head()

df_test_new['age'] = df_test_new['age'] - 3

bins = np.bincount(df_test_new['age'].values)
bins

df_test_new.to_csv('afad_test_balanced.csv', index=None)
