==> notebooks/08_california_housing-balanced.ipynb <==
import pandas as pd
import numpy as np

# Train

df = pd.read_csv('../data/processed/california_housing/californiahousing.csv')
df.head()

bins = np.bincount(df['response'].values)
bins

bins[1:]

lst = []
for idx, i in enumerate(bins[1:]):
    tmp = df[df['response'] == idx+1]
    lst.append(tmp.iloc[:1379, ])

df_new = pd.concat(lst)

df_new.head()

df_new['response'] = df_new['response'] - 1

bins = np.bincount(df_new['response'].values)
bins

## Split

indices = np.arange(df_new.shape[0])
labels = df_new['response'].values

from sklearn.model_selection import train_test_split


indices_train, indices_test, labels_train, labels_test = train_test_split(
    indices, labels, test_size=0.2, random_state=123, shuffle=True, stratify=labels)

indices_train, indices_valid, labels_train, labels_valid = train_test_split(
    indices_train, labels_train, test_size=0.05, random_state=123, shuffle=True, stratify=labels_train)

df_train = df_new.iloc[indices_train]
df_train.to_csv('../data/processed/california_housing/californiahousing_balanced_train.csv', index=None)

df_valid = df_new.iloc[indices_valid]
df_valid.to_csv('../data/processed/california_housing/californiahousing_balanced_valid.csv', index=None)

df_test = df_new.iloc[indices_test]
df_test.to_csv('../data/processed/california_housing/californiahousing_balanced_test.csv', index=None)
