==> notebooks/09_aes-balanced.ipynb <==
import pandas as pd
import numpy as np

# Train

df_train = pd.read_csv('aes_train.csv', index_col=0)
df_train.head()

bins = np.bincount(df_train['beauty_scores'].values)
bins

bins[1:-1]

lst = []
for idx, i in enumerate(bins[1:-1]):
    tmp = df_train[df_train['beauty_scores'] == idx+1]
    lst.append(tmp.iloc[:498, ])

df_train_new = pd.concat(lst)

df_train_new.head()

df_train_new['beauty_scores'] = df_train_new['beauty_scores'] - 1

bins = np.bincount(df_train_new['beauty_scores'].values)
bins

df_train_new.to_csv('aes_train_balanced.csv')

# Valid

df_valid = pd.read_csv('aes_valid.csv', index_col=0)
df_valid.head()

bins = np.bincount(df_valid['beauty_scores'].values)
bins

bins[1:-1]

lst = []
for idx, i in enumerate(bins[1:-1]):
    tmp = df_valid[df_valid['beauty_scores'] == idx+1]
    lst.append(tmp.iloc[:26, ])

df_valid_new = pd.concat(lst)

df_valid_new.head()

df_valid_new['beauty_scores'] = df_valid_new['beauty_scores'] - 1

bins = np.bincount(df_valid_new['beauty_scores'].values)
bins

df_valid_new.to_csv('morph2_valid_balanced.csv')

# Test

df_test = pd.read_csv('aes_test.csv', index_col=0)
df_test.head()

bins = np.bincount(df_test['beauty_scores'].values)
bins

bins[1:-1]

lst = []
for idx, i in enumerate(bins[1:-1]):
    tmp = df_test[df_test['beauty_scores'] == idx+1]
    lst.append(tmp.iloc[:96, ])

df_test_new = pd.concat(lst)

df_test_new.head()

df_test_new['beauty_scores'] = df_test_new['beauty_scores'] - 1

bins = np.bincount(df_test_new['beauty_scores'].values)
bins

df_test_new.to_csv('morph2_test_balanced.csv')
